+++ src/main/java/net/oneandone/stool/client/Client.java
@@ -64,7 +64,7 @@ public abstract class Client {
     public abstract void delete(String stage) throws IOException;
     public abstract Map<String, String> getValues(String stage) throws IOException;
     public abstract Map<String, String> setValues(String stage, Map<String, String> values) throws IOException;
-    public abstract List<String> history(String stage, boolean details, int max) throws IOException;
+    public abstract List<String> history(String stage) throws IOException;
     public abstract List<String> validate(String stage, boolean email, boolean repair) throws IOException;
     public abstract PodConfig podToken(String stage, int timeout) throws IOException;
     public abstract List<String> images(String stage) throws Exception;